Update: Use size_t Over unsigned
- mapping from Fortran to C is more explicit with use of size_t over
  unsigned

+++ include/stopwatch/stopwatch.h
@@ -1,6 +1,7 @@
 #ifndef STOPWATCH_STOPWATCH_H
 #define STOPWATCH_STOPWATCH_H
 
+#include <stddef.h>
 // Function status constants
 #define STOPWATCH_OK 0
 #define STOPWATCH_ERR (-1)
@@ -32,8 +33,8 @@ struct StopwatchMeasurementResult {
   long long total_event_values[STOPWATCH_MAX_EVENTS];
   long long total_times_called;
   char routine_name[NULL_TERM_MAX_ROUTINE_NAME_LEN];
-  unsigned int stack_depth;
-  unsigned int num_of_events;
+  size_t stack_depth;
+  size_t num_of_events;
   int event_names[STOPWATCH_MAX_EVENTS];
 };
 
@@ -44,7 +45,7 @@ struct StopwatchMeasurementResult {
 // Initializes the event timers. Currently the events that are measured are hard coded. This will also start the
 // monotonic measurement clock as currently it is assumed that consumers would immediately start the clock after
 // initializing the stopwatch structure.
-int stopwatch_init(const enum StopwatchEvents *events_to_add, unsigned int num_of_events);
+int stopwatch_init(const enum StopwatchEvents *events_to_add, size_t num_of_events);
 
 // Stops the monotonic event timers and cleans up resources used by the timer. Interestingly valgrind still reports a
 // memory leak with the PAPI specific resources
@@ -55,14 +56,14 @@ void stopwatch_destroy();
 // =====================================================================================================================
 
 // Records the current values on the monotonic event timers
-int stopwatch_record_start_measurements(int routine_call_num, const char *function_name, unsigned int stack_depth);
+int stopwatch_record_start_measurements(int routine_call_num, const char *function_name, size_t stack_depth);
 
 // Records the current values on the monotonic event timers. Will also perform a delta between the values recorded from
 // `stopwatch_record_start_measurements` and the current values to give a measurement on the profile on the procedure executing
 // between the calls of `stopwatch_record_start_measurements` and `stopwatch_record_end_measurements`
 int stopwatch_record_end_measurements(int routine_call_num);
 
-int stopwatch_get_measurement_results(unsigned int routine_call_num, struct StopwatchMeasurementResult *result);
+int stopwatch_get_measurement_results(size_t routine_call_num, struct StopwatchMeasurementResult *result);
 
 // Prints out the results
 void stopwatch_print_measurement_results(struct StopwatchMeasurementResult *result);
